Get nameservers in Windows OS
